Get account currency along with currency_id. Closes #3

+++ src/test/java/jdc/spring/accountwebservice/controller/AccountControllerTest.java
@@ -20,7 +20,7 @@ public class AccountControllerTest {
 
     @Test
     public void getAllAccounts() {
-        Long idValue = 2L;
+        int idValue = 2;
 
         accounts.setId(idValue);
